handson/01-ec2/requirements.txt
aws-cdk.aws-ec2==1.44.0
